move general settings to inner settings page
