+++ app/src/main/java/com/parishod/watomatic/fragment/SettingsFragment.java
@@ -3,21 +3,16 @@ package com.parishod.watomatic.fragment;
 import android.content.Intent;
 import android.os.Bundle;
 
-import androidx.preference.ListPreference;
 import androidx.preference.Preference;
 import androidx.preference.PreferenceFragmentCompat;
 import androidx.preference.SwitchPreference;
 
 import com.parishod.watomatic.R;
 import com.parishod.watomatic.activity.advancedsettings.AdvancedSettingsActivity;
-import com.parishod.watomatic.activity.customreplyeditor.CustomReplyEditorActivity;
-import com.parishod.watomatic.activity.main.MainActivity;
-import com.parishod.watomatic.model.preferences.PreferencesManager;
 import com.parishod.watomatic.model.utils.AutoStartHelper;
 import com.parishod.watomatic.model.utils.ServieUtils;
 
 public class SettingsFragment extends PreferenceFragmentCompat {
-    private ListPreference languagePref;
     private SwitchPreference showNotificationPref, foregroundServiceNotifPref;
     private Preference advancedPref;
     private Preference autoStartPref;
@@ -26,19 +21,6 @@ public class SettingsFragment extends PreferenceFragmentCompat {
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
         setPreferencesFromResource(R.xml.fragment_settings, rootKey);
 
-        languagePref = findPreference(getString(R.string.key_pref_app_language));
-        languagePref.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-            @Override
-            public boolean onPreferenceChange(Preference preference, Object newValue) {
-                String thisLangStr = PreferencesManager.getPreferencesInstance(getActivity()).getSelectedLanguageStr(null);
-                if(thisLangStr == null || !thisLangStr.equals(newValue)){
-                    //switch app language here
-                    //Should restart the app for language change to take into account
-                    restartApp();
-                }
-                return true;
-            }
-        });
 
         showNotificationPref = findPreference(getString(R.string.pref_show_notification_replied_msg));
         if (android.os.Build.VERSION.SDK_INT < android.os.Build.VERSION_CODES.M) {
@@ -78,15 +60,16 @@ public class SettingsFragment extends PreferenceFragmentCompat {
         });
     }
 
+    @Override
+    public void onResume() {
+        super.onResume();
+        getActivity().setTitle(R.string.settings);
+    }
+
     private void checkAutoStartPermission() {
         if(getActivity() != null) {
             AutoStartHelper.getInstance().getAutoStartPermission(getActivity());
         }
     }
 
-    private void restartApp() {
-        Intent intent = new Intent(getActivity(), MainActivity.class);
-        getActivity().startActivity(intent);
-        getActivity().finishAffinity();
-    }
 }
