+++ app/src/main/res/xml/fragment_general_settings.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <PreferenceCategory android:title="@string/preference_category_general_label">
+
+        <ListPreference android:title="@string/app_language"
+            app:icon="@drawable/ic_language"
+            android:key="@string/key_pref_app_language"
+            android:entries="@array/app_language"
+            android:entryValues="@array/app_language_value"
+            app:useSimpleSummaryProvider="true"/>
+
+        <CheckBoxPreference android:title="@string/add_attribution_label"
+            android:icon="@drawable/ic_attribution"
+            android:key="@string/pref_is_append_watomatic_attribution"
+            android:summary="@string/append_watomatic_arribution_checkbox_label"/>
+      
+    </PreferenceCategory>
+
+</PreferenceScreen>
