+++ app/src/main/res/xml/fragment_settings.xml
@@ -3,19 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <PreferenceCategory android:title="@string/preference_category_general_label">
-
-        <ListPreference android:title="@string/app_language"
-            app:icon="@drawable/ic_language"
-            android:key="@string/key_pref_app_language"
-            android:entries="@array/app_language"
-            android:entryValues="@array/app_language_value"
-            app:useSimpleSummaryProvider="true"/>
-
-        <CheckBoxPreference android:title="@string/add_attribution_label"
-            android:icon="@drawable/ic_attribution"
-            android:key="@string/pref_is_append_watomatic_attribution"
-            android:summary="@string/append_watomatic_arribution_checkbox_label"/>
-      
+        <Preference android:title="@string/preference_category_general_label"
+            android:fragment="com.parishod.watomatic.fragment.GeneralSettingsFragment"
+            />
     </PreferenceCategory>
 
     <PreferenceCategory android:title="@string/preference_category_notifications_label"
